refactor: 구입 금액 1000원 단위 검증 테스트 ValidatorTest로 이동
