+++ src/test/java/lotto/validation/ValidatorTest.java
@@ -1,0 +1,34 @@
+package lotto.validation;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class ValidatorTest {
+
+    private static final String ERROR_MESSAGE = "[ERROR]";
+
+    @Test
+    void 구입_금액이_1000원_단위이면_테스트에_성공한다() {
+        // given
+        Validator validator = new Validator();
+        int purchaseAmount = 3000;
+
+        // when & then
+        assertThatCode(() -> validator.validatePurchaseAmount(purchaseAmount))
+                .doesNotThrowAnyException();
+    }
+
+    @Test
+    void 구입_금액이_1000원_단위가_아니면_예외를_던진다() {
+        // given
+        Validator validator = new Validator();
+        int purchaseAmount = 3500;
+
+        // when & then
+        assertThatThrownBy(() -> validator.validatePurchaseAmount(purchaseAmount))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining(ERROR_MESSAGE);
+    }
+}
